internal/socket: new package
This change introduces a package that provides a portable interface
for the manipulation of socket; a service access point for transport-
or network-layer connection termination point.

The motivation of this change is not only for fixing broken packages
in golang.org/x/net repository due to some change in Go 1.9, but for
providing a low-level portable interface that can be used as a partial
replacement for the frozen syscall package in future.

The syscall package provides some part of the manipulation for socket
but is incomplete: for example, some system calls are not implemented
due to the mismatch between the exposed types and latest kernels; the
sendmsg and recvmsg system calls don't provide scatter-gather I/O for
a single end point which is useful for performance improvement when
using datagram protocols; the {get,set}sockopt system calls accept
only options known by the syscall package.

In order to fix the issues in syscall pacakge without causing new
problems, the package introduced by this change is considered to be an
internal package which helps to construct, maintain exposed API.

As a first step, the package provides a few APIs as minimum
requirements for fixing golang.org/x/net/{icmp,ipv4,ipv6} packages.

Updates #19051.

Change-Id: I63dc4d8267dc49bf6f0a0c4c66b7a86b1922ccfb