Update patch set 1
Patch Set 1: Code-Review-1

I'd rather not do this as it opens the possibility of doing O(n) work for a single insert (if we have to skip a lot of buckets before we find an unevacuated one).  It's not a lot of work to check the evacuation status of a bucket, but still.

I think I like the O(1) buckets instead of 1 bucket CL better.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1