runtime: time is correctly updated on Wine
Implemented low-level time system for windows on hardware (software),
which does not support memory mapped _KSYSTEM_TIME page.

In particular this problem exists on wine where _KSYSTEM_TIME
only contains time at the start, and is never modified.

On start we try to detect that the page is not updated. If it's so
we fallback to a monotonic timer based on QueryPerformanceCounter family
of syscalls:
https://msdn.microsoft.com/en-us/library/windows/desktop/dn553408(v=vs.85).aspx

This fixes #18537

Change-Id: I269d22467ed9b0afb62056974d23e731b80c83ed