runtime: stack traces of endless recursion now print top and bottom
This CL prints out the top and bottom 50 frames per stack trace, instead
of massive stack traces. If the number of frames <= 100
(_TracebackMaxFrames), we'll just print all the frames out.

While here, also fix a bug that resulted from CL 37222, in which we
no longer accounted for decrementing requested frame skips, and assumed
that when printing, that skip would always be 0, so instead added a
precondition that we'll only print if skip <= 0.

Fixes #7181.
Fixes #24628.

Change-Id: Ie31ec6413fdfbe43827b254fef7d99ea26a5277f