Update patch set 1
Patch Set 1:

Maybe the bug is easy to fix instead.

I'm also concerned that this documents the current status but will go out of date as browsers change.

Maybe softer wording like "The trace viewer comes out of the Chrome/Chromium project and is only actively tested in that browser." ?

Patch-set: 1
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>