Update patch set 1
Patch Set 1:

Seems like a good idea.  I was wondering a bit about NilChecks, and could one of those sink, but not a subsequent op that it was supposed to be guarding against.  Or perhaps, a NilCheck for effect (because of pointer arithmetic) could get sunk into a conditionally executed block?

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>