Update patch set 3
Patch Set 3:

> > I was wondering a bit about NilChecks, and could one of those sink, but not a subsequent op that it was supposed to be guarding against.  Or perhaps, a NilCheck for effect (because of pointer arithmetic) could get sunk into a conditionally executed block?
> 
> The second one definitely seems like a problem. Not sure about the first. Will exclude NilChecks explicitly.

Although I wonder whether this could also be a problem for loads that could fault, and whose nil check was removed in favor of an implicit (page fault) nil check.

If so, we could restrict this to only apply to SP (and SB?) offptrs, which should never fault, and which are the cases that this optimization is intended to help with.

Patch-set: 3
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>