runtime: augment inuse heap profile with collected garbage
DO NOT SUBMIT. This is an experiment.

This augments the usual inuse heap profile to show memory collected by
the last GC cycle in addition to memory retained by it. This is a more
holistic "memory" profile of the process that reflects its total
memory footprint, rather than just a "heap" profile. It shows,
directly in the default profile, the missing half of the memory
footprint that often confuses new users.

This is a prototype, so the implementation simply uses the profile
label mechanism to label the original samples "gc:live" and to label
the new dead samples "gc:dead". This isn't very visually compelling in
the pprof output, but it works and lets you filter on these tags.

Updates #13463.

Change-Id: I989c9ce52ee9344f827ddce55adf2529f2e45b21