Update patch set 6
Patch Set 6:

> But then your change will never work on real Windows. I thought
 > anything that works on Wine should work on Windows and vice versa.
 
It is not supposed to run on real windows, just because real windows have to have mapped page with updated timer structures.
Windows just absolutely have to have it.

I will cook up a new patch which will probably work on windows, but timing still will not be precise because of 64 bit(!!!) multiplication overflow and fixed integer devision.
 
 > Your are adding too many lines just to support Wine. You should
 > make code simpler. We need to support this code when it breaks.

Wine does not support that it has to be fixed somehow. I proposed the solution, it may look a bit large, but it doesn't, it just a wrapper on top of a single syscall. Patch was like 3 times less before you moved memory reading and helper functions into asm, btw.

I do not understand your point. Wine is not Windows, it tries its best to look like it, but it fails in many places, which have to be fixed, so I fixed it in time code - everywhere where I see the problem.

 > And you are making Windows code slower in the process. I can see
 > all speed ups gained by https://go-review.googlesource.com/36428
 > are gone now.
 
Frankly saying, for real windows my code does nothing but a single function call on top of asm, if you do care THAT much about time of the function call, it is not a time system you should optimize but go compiler.

And I personally never considered a time of the function call a somewhat meaningful benchmark of the language as well as a problem.

Do you object against supporting Wine in golang and the reason for that is few additional nanoseconds in time.Now() call? Was it EVER a problem?

Patch-set: 6
Reviewer: Gerrit User 17055 <17055@62eb7196-b449-3ce5-99f1-c037f21e1705>