Update patch set 6
Patch Set 6:

Evgenij,

In Gerrit, a change is almost always a single commit and the primary key of a change is the "Change-Id". That's what Gerrit uses to know that 10 different "versions" of a commit are all related and are "CL 35710".

So just squash your changes back into a single commit and make sure the "Change-Id: Ixxxxxxxxxx" line is the same as the first version here.

Patch-set: 6
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>