Update patch set 7
Patch Set 7:

It's been a while since I looked at this, but is there an actual reason that file names are stuff into symbols? There's a comment somewhere about how they should just be stored as strings (this would probably motivate adding a string table to the object file format, but that's not that hard, I've done it at least twice in local experiments).

Patch-set: 7
CC: Gerrit User 5153 <5153@62eb7196-b449-3ce5-99f1-c037f21e1705>