Update patch set 7
Patch Set 7:

Changes to the object format are on hold. This is fine for now.

They're stuffed into the symbol table because they were in the C version. In the C version they were likely that way because the symbol table was the only string-keyed hash table available.

Patch-set: 7
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>