testing: fix Benchmark() to start at 1 iteration, not 100
The run1 call removed in golang.org/cl/36990 was necessary to
initialize the duration of the benchmark. With it gone, the math in
launch() starts from 100. This doesn't work out well for second-long
benchmark methods.

Instead of putting the run1 call back, fix the math to behave better
when run from scratch, and hopefully make it more readable.

Updates #18815

Change-Id: I461f3466c805d0c61124a2974662f7ad45335794