Update patch set 7
Patch Set 7:

I would rather not drop QPC frequency from the table just because it happens to be constant 10e6 for wine, this is not documented and will explode if something changes. It will also break if real hardware happens to miss the check which was the way Alex tested this patchset, in fact, this is the only way to test it, since there are too many failed test cases on wine.

Patch-set: 7
Reviewer: Gerrit User 17055 <17055@62eb7196-b449-3ce5-99f1-c037f21e1705>