Update patch set 2
Patch Set 2:

On further reflection, I prefer CL 37011. The reason is that CL 37011 should not increase worst case time. Instead of sometimes evacuating 1 and sometimes evacuating 2, we already evacuate 2. (The overhead of searching for the next bucket to evacuating is vanishingly small.)

We could also potentially do both. And we could get fancier here and tailor the aggressiveness of the map work to the sizes of the buckets.

But my inclination is:

* Do CL 37011 now.
* Then do evacuate-on-read.
* Wait for a release and see how the above fairs in production in practice.
* If all goes well, I will re-evaluate this CL for 1.10.

Patch-set: 2
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>