Update patch set 8
Patch Set 8:

> Uploaded patch set 8.

This latest patch is following a clarifying conversation with @adg about the existing terminology. 

The document now consistently uses:

Send: to describe the act of sending a change to the code review server
Apply: to describe the act of applying a change to the master branch
git mail: as the action to send a change for review
Submit: as the action to apply a change
commit: always describes a git commit
change: always describes a gerrit change

The general terms "send" and "apply" are used to describe the broad operation of what is happening. These terms are approachable for newcomers as regardless of their existing workflow experience they are likely familiar with the concept of sending and applying changes.

The specific terms "mail" & "submit" are used when describing what to do/run.

Patch-set: 8
Reviewer: Gerrit User 14840 <14840@62eb7196-b449-3ce5-99f1-c037f21e1705>