Update patch set 8
Patch Set 8:

(7 comments)

I've gone through and standardized all of the new ids created by this change. I've ensured that all of the existing ids are left untouched. In places where an id should be renamed as the old id name is no longer applicable I've created an empty span with the old id adjacent to the renamed one.

Patch-set: 8