Update patch set 10
Patch Set 10:

> (1 comment)

Submit isn't specific to Gerrit, but the usage of this term is AFAIK unique to Gerrit. This article refers to it as a less common synonym for "commit" but that's not how we use it anyway.

In our workflow we use "commit", "mail" and "submit". These terms are very confusing as we've redefined "submit" from how it is used in most other VSCs. The only VCS I could find that uses "submit" in the way Gerrit does is Perforce which is likely why the term is used in Gerrit.

There is also the additional challenge of trying to use general terms for two very different types of systems, distributed VCS (Git, Mercurial) and centralized VCS (Subversion, Perforce, CVS)

The term "apply" is broadly used to refer to applying one set of changes onto a code base. You "apply a patch"; You "apply a diff". Patch, Diff, Subversion, CVS, Git all use this term in their documentation to refer to this. This is the exact term used in the git documentation to describe what a "git cherry-pick" does. https://git-scm.com/docs/git-cherry-pick . In all of them they use the term "apply" in the same way I'm proposing here.

If we don't like "apply" then as as an alternative I do like the term "promote" which is listed in this article. It's even more descriptive as you are not only applying the change but also applying it to a more restricted branch (master).

Patch-set: 10
Reviewer: Gerrit User 14840 <14840@62eb7196-b449-3ce5-99f1-c037f21e1705>