cmd/compile: put spills in better places
Previously we always issued a spill right after the op
that was being spilled.  This CL pushes spills father away
from the generator, hopefully pushing them into unlikely branches.
For example:

  x = ...
  if unlikely {
    call ...
  }
  ... use x ...

Used to compile to

  x = ...
  spill x
  if unlikely {
    call ...
    restore x
  }

It now compiles to

  x = ...
  if unlikely {
    spill x
    call ...
    restore x
  }

This is particularly useful for code which appends, as the only
call is an unlikely call to growslice.

The basic algorithm is to put the spill at the lowest point in
the dominator tree that:
 1) dominates all restores
 2) has the value being spilled in a register
 3) has a loop depth no deeper than the value being spilled

TODO: this code is way too slow.
 1) We satisfy 1,2,3 above iteratively. Do that faster.
 2) Splicing spill ops in to the right spot in the block schedule is too slow.

I had to delete most of David's code which pushed spills out of loops.
I suspect this CL subsumes most of the cases that that code handles.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: Ia5116e53a3aeb025ef350ffc51c14ae5cc17871c