cmd/compile: allow composite literal structs with _ fields
Given code such as

type T struct {
  _ string
}

func f() {
  var x = T{"space"}
  // ...
}

the compiler rewrote the 'var x' line as

var x T
x._ = "space"

The compiler then rejected the assignment to
a blank field, thus rejecting valid code.

The only way such an assignment is legal
is to use a composite literal in which
the struct fields are provided implicitly.
When we encounter such struct fields,
mark them as allowing assignment to a blank field,
and check for the exception at the right moment.

Fixes #19482

Change-Id: I594476171d15e6e8ecc6a1749e3859157fe2c929