Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We do have tests for stack tracebacks, e.g., TestPanicTraceback.  Should we have a test here?

We even have a test for GOTRACEBACK=crash (TestCrashDumpsAllThreads), but I couldn't figure out how to write a reliable automated test for this because it requires getting a thread on to the system stack and keeping it there for a long time (or at least a lot of the time). That is, by design, hard for user code to do. In the test case I wrote for the issue, I had to carefully align timing of the GC and rely on the fact that typedslicecopy happens to run on the system stack when GC is active. And since we have to build an external test binary so we have something to signal, we don't have a way to export runtime internals to it.

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>