Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Can't hurt, but I don't think this is the fix.
> 
> Some the failures look like they get the backwards results for the two connections (they get each other's addresses), which seems unrelated to load or a 1 second timeout.

I think they might be related. We start the first request and block until the connection is accepted, but that doesn't mean that the handler func has actually started running for it. Under high load, it's possible that we then start the second request, start the handler func for the second request, then time out, leading to reversed results.

Huh. Is that the kind of thing you were asking for in the issue when you asked for a race? :P

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>