*.bash: use clearer version of string equality
POSIX Shell only supports = to compare variables inside '[' tests. But
this is Bash, where == is an alias for =. Since we write in Bash and ==
is a clearer form, use it everywhere.

It was already the more common form in all the bash scripts, but quite a
few mixed the two forms.

Change-Id: I38fa7a5a90658dc51acc2acd143049e510424ed8