cmd/compile/internal/gc: remove unused exportsize variable
In Go 1.7 and earlier, gc.exportsize tracked the number of bytes
written through exportf. With the removal of the old exporter in Go 1.8
exportf is only used for printing the build id, and the header and
trailer of the binary export format. The size of the export data is
now returned directly from the exporter and exportsize is never
referenced. Remove it.

Change-Id: Id301144b3c26c9004c722d0c55c45b0e0801a88c