Update patch set 5
Patch Set 5:

The new message seems worse to me.

"want 2 values, got 1"

Who wants 2 values? The LHS? The RHS? Does it mean there should have be two variables on the LHS but we only had one? Does it mean there are two variables on the LHS but the function only returns one? I have no idea.

With the old "2 = 1", it is immediately obvious to me (at least) that there are two variables on the LHS but the RHS returns only 1.

Patch-set: 5
CC: Gerrit User 5300 <5300@62eb7196-b449-3ce5-99f1-c037f21e1705>