Update patch set 9
Patch Set 9:

> Commit message should reference the git hash of the source code.

Done (I think?  My git-fu is weak.)

> This is going into sync instead of an internal vendor package though?

Yes: it's adding syncmap.Map as sync.Map, so it's not a clean vendoring.
(I don't think it can be vendored cleanly even with aliases, since it uses sync.Map.)

If you'd rather we vendor it in as some other subpackage of sync, I'm open to suggestions.

Patch-set: 9