encoding/gob: Speedup map decoding by reducing the allocations.
The time to decode a simple string to int map is reduced by 50%. This
is still 60% slower than using two arrays for keys and values and
recreating a map from them.

The improvements are achieved in two ways:
1) Allow to hint the hashmap about the incoming size growth. This way
hashmap can reallocate the buckets to the size that will accomodate
the growth in one operation, instead of performing multiple
allocations on individual inserts.
2) In encoding/gob/decode.go decodeMap allocate keyInstr and
elemInstr only once and pass it to decodeIntoValue, instead
of allocating a new instance on every loop cycle.

Fixes #19525

Change-Id: Iec5fa4530de76f0a70da5de8a129a567b4aa096e