Update patch set 1
Patch Set 1:

Small decimal numbers are probably very common. I can see that it makes sense to specialize for them. Maybe we should just look at 2, 4, 6, and 8 digit numbers and and have special-case code for them, and simplify the rest instead? Using the table with 100 values that we have now would allow these 4 cases to be handled with unrolled loops.

And we should use the same code (or an identical copy) in fmt.

But again, lets' not get carried away here. It's easy to spend a lot of time improving this package with marginal impact on real programs.

I'm ok with more experimenting, but let's not go forward with every single incremental improvement that adds complexity. Thanks.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>