Update patch set 1
Patch Set 1:

"And we should use the same code (or an identical copy) in fmt."
I am not so sure about adding the lookup table approach also to fmt.
We would need to add another string array in fmt which adds to every binary
and fmt does not get the benefit of fewer allocations because it always formats into a preallocated buffer anyway. The overall effect on fmt is also smaller because fmt does more work outside of integer formatting than strconv. When i tested this in the past it seemed not worth it. But we can revisit again once a new strconv is committed.

However introducing separate 32bit and 64bit paths seems like a good idea for 32bit platforms at least. I have made a CL and i am benchmarking that currently.

Patch-set: 1