cmd/compile/internal/syntax: track precise [begin, end[ node positions
This change introduces exact begin/end positions for all nodes
such that the open-ended source interval [n.Begin(), n.End()[
corresponds to the syntactic construct represented by a node.
Some of these positions are explicitly recorded in nodes, and
some are computed on demand.

As before, each node has a "regular" position n.Pos() which
corresponds to the position of the left-most token of a construct
uniquely identifying the construct; this position is the one that
should be used for error reporting.

The change removes the node.pos field that was present in all nodes
in favor of node-specific position information.

Unfortunately it also increases the amount of memory used to parse
the std library from ~389MB to ~452MB; a 16% increase
(run: go test -run StdLib -fast).

This is only affecting the new syntax tree which is a small portion
of the memory allocated for a full compilation. (I will try to reduce
this number again, going forward).

There's also a minor performance penalty for parsing only, but it's
too small to affect the compiler speed in any significant way.

Additional tests have been added to verify correctness of the
[begin, end[ interval for all nodes (with a couple of minor exceptions
that need some fine-tuning). We expect to improve this code over time.

Change-Id: I3e4c75dad6c79c3ddf7e18560dcdc46d77dbfbe3