Update patch set 2
Patch Set 2:

Instead of setting minimum requirements for AMD64 to include POPCNT we could explore if its warranted to add a new (sub) ARCH target SSE4 (SSE4.2+POPCNT...) so everything starting with Nehalem and Bulldozer is covered. If that is a good starting point and if there are many optimization opportunities that would benefit from that step could surveyed and discussed on the issue tracker. Also seems a good time to spell out the deprecation policy in case amd64 requirements are increased and announce 1-2 releases in advance https://github.com/golang/go/issues/19075. Just bumping AMD64 requirements because of POPCNT seems not enough in my opinion to drop support for all of e.g. core 2 yet.

Patch-set: 2
CC: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>